requirements.txt
customtkinter==5.2.2
yt-dlp==2024.10.22
Pillow==10.4.0
requests==2.32.3